Changes in regards to keeping artifacts in memory, weka call
There were several directions for this modifications: 
1. when loading data (students, courses, etc), keep them in memory.
InputFileHandlers has maps with students (id <-> student), courses
(similar), instructors (similar). These are meant to find easily the
artifact with a certain id. 
2. change the weka code. Basically we need associations between courses,
so the structure to be analyzed has lines with students courses (each
student one line with the courses taken). The table with records is not
proper for this analysis. Different results than the ones in assignment,
on the other one the ones in assignment don't really make sense.